Added async emit approach instead of sync
This was the better way to handle the retry connections problems with
concurrent send operations.
Also the api was better for the way we handle this send operations and
the error handling will be done inside one single place.

A new sync api can be done will be added in the next commits

Also I remove the build tag for integration tests and start using a
simple flag for this operations and skip the integrations if the test
command didn't pass the `-integration` flag.