Use the default linter plus some optional linters
The golangci-lint adds more linters with some frequency.
Some of them are incompatible with others and
some make sense only in specific cases.

Instead of enabling all and block the linters a better approach is to enable
the defaults and a list of linters that make sense for this project